+++ src/index.ts
@@ -9,6 +9,7 @@ export { TraversingComponent } from './traversing';
 export { EditView } from './views/edit';
 export { LoginView } from './views/login';
 export { SearchView } from './views/search';
+export { SitemapView } from './views/sitemap';
 export { ViewView } from './views/view';
 export { Breadcrumbs } from './components/breadcrumbs';
 export { Comments, Comment, CommentAdd } from './components/comments';