+++ src/views/sitemap.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'plone-login',
+  template: `<h1>Sitemap</h1>
+  <plone-navigation root="/" [depth]="7"></plone-navigation>`
+})
+export class SitemapView {}